Checking whether entered page number is correct
Checking if the entered page number lies between 1 and the last page number for the selected pdf.
If not, then the user is prompted to enter a valid page number.
The function valid_input_checker is given an additional parameter 'pdfReader_File' which is critical for finding the page count of a pdf.